+++ Cargo.toml
@@ -4,7 +4,7 @@ resolver = "2"
 
 [profile.dev]
 opt-level = 3
-debug = 1              # change to 0 or 2 for more or less debug info
+debug = 2              # change to 0 or 2 for more or less debug info
 overflow-checks = true
 incremental = true
