+++ hashes/zkevm/Cargo.toml
@@ -12,9 +12,13 @@ itertools = "0.11"
 lazy_static = "1.4"
 log = "0.4"
 num-bigint = { version = "0.4" }
-halo2-base = { path = "../../halo2-base", default-features = false }
+halo2-base = { path = "../../halo2-base", default-features = false, features = [
+    "test-utils",
+] }
 rayon = "1.7"
 sha3 = "0.10.8"
+pse-poseidon = { git = "https://github.com/axiom-crypto/pse-poseidon.git" }
+getset = "0.1.2"
 
 [dev-dependencies]
 criterion = "0.3"
