+++ hashes/zkevm/src/keccak/coprocessor/circuit/mod.rs
@@ -1,0 +1,3 @@
+pub mod leaf;
+#[cfg(test)]
+mod tests;
