+++ hashes/zkevm/src/keccak/coprocessor/circuit/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+pub mod leaf;
