+++ hashes/zkevm/src/keccak/coprocessor/param.rs
@@ -1,0 +1,12 @@
+pub const OUTPUT_NUM_COL_COMMIT: usize = 1;
+pub const OUTPUT_NUM_COL_RAW: usize = 3;
+pub const OUTPUT_COL_IDX_COMMIT: usize = 0;
+pub const OUTPUT_COL_IDX_KEY: usize = 0;
+pub const OUTPUT_COL_IDX_HASH_LO: usize = 1;
+pub const OUTPUT_COL_IDX_HASH_HI: usize = 2;
+
+pub const POSEIDON_T: usize = 3;
+pub const POSEIDON_RATE: usize = 2;
+pub const POSEIDON_R_F: usize = 8;
+pub const POSEIDON_R_P: usize = 57;
+pub const POSEIDON_SECURE_MDS: usize = 0;
