+++ hashes/zkevm/src/keccak/coprocessor/tests/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(test)]
+mod output;
