+++ hashes/zkevm/src/keccak/vanilla/param.rs
@@ -1,5 +1,5 @@
 #![allow(dead_code)]
-pub(crate) const MAX_DEGREE: usize = 4;
+pub(crate) const MAX_DEGREE: usize = 3;
 pub(crate) const ABSORB_LOOKUP_RANGE: usize = 3;
 pub(crate) const THETA_C_LOOKUP_RANGE: usize = 6;
 pub(crate) const RHO_PI_LOOKUP_RANGE: usize = 4;
